ble/requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
bidict==0.21.4
bleak==0.14.1
certifi==2021.10.8
charset-normalizer==2.0.10
dbus-next==0.2.3
frozenlist==1.3.0
idna==3.3
isort==5.10.1
jedi==0.18.1
lazy-object-proxy==1.7.1
mccabe==0.6.1
multidict==6.0.2
parso==0.8.3
platformdirs==2.4.1
pluggy==1.0.0
pylint==2.12.2
pyls==0.1.6
python-engineio==4.3.1
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.3.3
python-socketio==5.5.1
requests==2.27.1
toml==0.10.2
ujson==5.1.0
urllib3==1.26.8
websocket-client==1.2.3
wrapt==1.13.3
yarl==1.7.2